compose_response() has been updated. On add2DB(), a feature that authenticate sms before processing it has been added. All sms's that are not in the format we want are not processed.

+++ Server_side_BP_system.py
@@ -59,6 +59,8 @@ def add2DB(data=[]):
     #create cursor to work with db
     Cursor = dbconnector.cursor()
 
+   
+
     try:
         
         for i in range(1, len(data)):
@@ -70,19 +72,24 @@ def add2DB(data=[]):
                 splitted = current_str[5].split("\r",2)
                 tstamp, payload = splitted[0], splitted[1]
                 accession_No = payload.split("|")
-                #print(accession_No[0])
-                #print("TIMESTAMP")
-                print (tstamp)
-                SQL = "INSERT INTO sms_RX_Q(sms_No, Rec, phone_No, time_stamp, payload,accession_No) values(?,?,?,?,?,?)"
-                data = (current_str[0], strip_quotes(current_str[1]), strip_quotes(current_str[2]), strip_quotes(current_str[4] + tstamp), payload, accession_No[0])
-                delete_all_GSM()
-                try:
-                    Cursor.execute(SQL, data)
-                    dbconnector.commit()
-                    #print("=====")
-                    
-                except Error as er:
-                    print("An error occured:", er.args[0])
+
+                if len(accession_No) > 1:
+                    #print(accession_No[0])
+                    #print("TIMESTAMP")
+                    print (tstamp)
+                    SQL = "INSERT INTO sms_RX_Q(sms_No, Rec, phone_No, time_stamp, payload,accession_No) values(?,?,?,?,?,?)"
+                    data = (current_str[0], strip_quotes(current_str[1]), strip_quotes(current_str[2]), strip_quotes(current_str[4] + tstamp), payload, accession_No[0])
+                    delete_all_GSM()
+                    try:
+                        Cursor.execute(SQL, data)
+                        dbconnector.commit()
+                        #print("=====")
+                        
+                    except Error as er:
+                        print("An error occured:", er.args[0])
+
+                else:
+                    pass
 
     except IndexError:
         pass
@@ -101,60 +108,65 @@ def process_sms():
 
     #moving records from sms_RX_Q table
     for row in rows:
-        time_stamp = row[0]     
-        phone_No = row[1]       
-        payload = row[2]        
-        payload = row[2].split("|")             
-        accession_No = payload[0].split("-")
+        timestamp = row[0]
+        phone =  row[1]
+        payload = row[2]
 
-        mac_address = accession_No[1]
-        accession_No = payload[0]
-        natID = payload[1]
-        first_name = payload[2]     
-        last_name = payload[3]
-        sex = payload[4]            
-        DOB = payload[5]
-        sys_mmHg = payload[6]       
-        dia_mmHg = payload[7]
-        
-        
-        try:
-            # moving records from sms_TX_Q table and insert them into other tables in db
-            # delete processed records in sms_TX_Q
+        if len(payload) > 0:
+            splitedPayload = payload.split("|")
+
+            accession_No = splitedPayload[0]
+            splited_accession_No = accession_No.split("-")
+
+            mac = splited_accession_No[1]
             
-            if  len(payload) > 0 :
+            natID = splitedPayload[1]
+            fName = splitedPayload[2]
+            lName = splitedPayload[3]
+            gender = splitedPayload[4]
+            DOB = splitedPayload[5]
+            BPsys = splitedPayload[6]
+            BPdia = splitedPayload[7]
+
+            try:
+                # moving records from sms_RX_Q table and insert them into other tables in db
+                # delete processed records in sms_RX_Q
 
                 #insert into demographic table
                 Cursor.execute("INSERT INTO demographic(ID_No, first_name, last_name, DOB, sex,accession_No) VALUES(?, ?, ?, ?, ?, ?)",
-                              (natID, first_name, last_name, DOB,sex, accession_No))
+                                (natID, fName, lName, DOB,gender, accession_No))
                 dbconnector.commit()
 
-                
+                    
                 #insert into Vitals table
                 Cursor.execute ("INSERT INTO Vitals(ID_No, time_stamp,sys_mmHg, dia_mmHg,accession_No) VALUES (?, ?,?, ?, ? )",
-                                (natID, time_stamp, sys_mmHg, dia_mmHg,accession_No))
+                                (natID, timestamp, BPsys, BPdia,accession_No))
                 dbconnector.commit()
 
                 #insert into SIM table
                 Cursor.execute ("INSERT INTO SIM(phone_No, mac_address, accession_No) VALUES (?, ?, ?)",
-                                (phone_No, mac_address, accession_No))
+                                (phone, mac, accession_No))
                 dbconnector.commit()
 
                 #insert into Nodes table
                 Cursor.execute ("INSERT INTO Nodes(mac_address, timestamp) VALUES (?, ? )",
-                                (mac_address, time_stamp))
+                                (mac, timestamp))
                 dbconnector.commit()
-                
+                    
                 #delete processed sms
-                Cursor.execute ("DELETE FROM sms_RX_Q WHERE time_stamp = \""+time_stamp+"\"")
+                Cursor.execute ("DELETE FROM sms_RX_Q WHERE time_stamp = \""+timestamp+"\"")
                 dbconnector.commit()
 
-        except sqlite3.IntegrityError as e:
+            except sqlite3.IntegrityError as e:
                 print('sqlite error: ', e.args[0])
-
                 
+
+        else:
+            process_sms()
+
 
 #======================================================================== COMPOSE RESPONSE =======================================================================================
+
 
 def compose_response():
     
@@ -163,92 +175,93 @@ def compose_response():
 
     #retrieve natID in first row
     Cursor = dbconnector.cursor()
-    Cursor.execute("SELECT ID_No FROM Vitals LIMIT 0,1")
+    Cursor.execute("SELECT ID_No,accession_No FROM Vitals LIMIT 0,1")
     rows = Cursor.fetchall()
 
     for row in rows:
-        natID = row[0]
-        #print(natID)
+        if len(row[0]) > 1:
+            natID = row[0]
+            accession_No1 = row[1]
+        else:
+            compose_response()
 
     Cursor = dbconnector.cursor()
-    Cursor.execute("SELECT sys_mmHg, dia_mmHg, accession_No FROM Vitals WHERE ID_No = \""+natID+"\" ORDER BY time_stamp DESC LIMIT 0,1")
+    Cursor.execute("SELECT sys_mmHg, dia_mmHg, accession_No FROM Vitals WHERE ID_No = \""+natID+"\" ORDER BY time_stamp LIMIT 0,1")
     rows = Cursor.fetchall()
 
     recommendation = ""
-    name = ""
+    comment = ""
+    reply = ""
     response = ""
+
     for row in rows:
-        row1_sys_mmHg = row[0]
-        row1_dia_mmHg = row[1]
+        current_BPsys = row[0]
+        current_BPdia = row[1]
         accession_No = row[2]
-        
-        #print(row1_sys_mmHg)
-        #print(row1_dia_mmHg)
 
-    Cursor = dbconnector.cursor()
-    Cursor.execute("SELECT sys_mmHg, dia_mmHg, accession_No FROM Vitals WHERE ID_No = \""+natID+"\" ORDER BY time_stamp DESC LIMIT 1,1")
-    rows = Cursor.fetchall()
-    for row in rows:
-        row2_sys_mmHg = row[0]
-        row2_dia_mmHg = row[1]
-
-        #print (row2_sys_mmHg)
-        #print (row2_dia_mmHg)
-
-    # composing a recommandation base on the BP range
-
-    if (row1_sys_mmHg in range (100, 140)) and (row1_dia_mmHg in range (50, 90)):
+    #Response
+            
+    if (current_BPsys in range (100, 140)) and (current_BPdia in range (50, 90)):
         recommendation = "Your Blood pressure is normal."
         print (recommendation)
 
-    elif (row1_sys_mmHg in range (141, 160)) and (row1_dia_mmHg in range (91, 100)):
+    elif (current_BPsys in range (141, 160)) and (current_BPdia in range (91, 100)):
         recommendation = "See clinician within a month."
         print (recommendation)
 
-    elif (row1_sys_mmHg > 179) and (row1_dia_mmHg > 110):
+    elif (current_BPsys > 179) and (current_BPdia > 110):
         recommendation = "See clinician within a week."
         print (recommendation)
 
     else:
         pass
 
+    
+    #Comparison
+    Cursor = dbconnector.cursor()
+    Cursor.execute("SELECT sys_mmHg, dia_mmHg, accession_No FROM Vitals WHERE ID_No = \""+natID+"\" ORDER BY time_stamp LIMIT 1,1")
+    rows = Cursor.fetchall()
+    for row in rows:
+        if len(str(row[0])) < 1  or len(str(row[1])) < 1:
+            pass
 
-    # Comparering current result with previous result
+        else:
+            previous_BPsys = row[0]
+            previous_BPdia = row[1]
 
-    if (row2_sys_mmHg > row1_sys_mmHg) and (row2_dia_mmHg > row1_dia_mmHg):
-        comment = "(Your BP is improving)"
-        print (comment)
 
-    else:
-        comment = "(Your BP is not improving)"
-        print (comment)
-        
-        
-        #accessing name of a person in db using accession_No and add it to the response
-        #used accession_No accessed from Vitals to access first_name and last_name in  demographic
-        Cursor = dbconnector.cursor()
-        Cursor.execute("SELECT first_name, last_name FROM demographic WHERE accession_No = \""+accession_No+"\"")
-        rows = Cursor.fetchall()
+            if (previous_BPsys > current_BPsys) and (previous_BPdia > current_BPdia):
+                comment = "After comparing current with previous BP, your BP has improved."
+                print (comment)
 
-        for row in rows:
-            first_name = row[0]
-            last_name = row[1]
+            elif (previous_BPsys < current_BPsys) and (previous_BPdia < current_BPdia):
+                comment = "After comparing with current and previous BP, your BP has not improved."
+                print (comment)
 
-            name = first_name + " " + last_name + ", " + "your previous BP; " + str(row1_sys_mmHg) + "/" + str(row1_dia_mmHg) + "," + "your current BP; " + str(row2_sys_mmHg) + "/" + str(row2_dia_mmHg)
-        
-        response = accession_No + "|" + name + " " + recommendation + comment    
-        #print(name)
-        #compose response
-        #print(response)
-        #print (recommendation)
+            else:
+                pass
 
-        try:
-            if len(response) > 0:
-                Cursor.execute("INSERT INTO sms_TX_Q(accession_No, response) VALUES(?, ?)",
-                            (accession_No,response))
-                dbconnector.commit()
+
+    #accessing name of a person in db using accession_No and add it to the response
+    #used accession_No accessed from Vitals to access first_name and last_name in  demographic
+    Cursor = dbconnector.cursor()
+    Cursor.execute("SELECT first_name, last_name FROM demographic WHERE accession_No = \""+accession_No1+"\"")
+    rows = Cursor.fetchall()
+
+    for row in rows:
+        fName = row[0]
+        lName = row[1]
+
+    reply = fName + " " + lName + ", " + "your current BP is " + str(current_BPsys) + "/" + str(current_BPdia) + "," + "your previous BP is " + str(previous_BPsys) + "/" + str(previous_BPdia) +". "
             
-        except Error as er:
+    response = accession_No1 + "|" + reply + " " + recommendation + comment    
+
+    try:
+        Cursor.execute("INSERT INTO sms_TX_Q(accession_No, response) VALUES(?, ?)",
+                    (accession_No,response))
+        dbconnector.commit()
+                
+    except Error as er:
             print (" An error occured", er.args[0])
 
 
@@ -304,7 +317,7 @@ def send_response():
                 print ("Nothing to send")
 
 
-#================================================================================= LOOP ==============================================================================
+#================================================================================= LOOP ==================================================================================
 
 
 while(True):
